% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getMeasures}
\alias{getMeasures}
\title{Get common co-occurrence measures.}
\usage{
getMeasures(
  data,
  lemma_names = c("lemma_1", "lemma_2"),
  slot_names = c("f", "s"),
  doc_id = "doc",
  measures = list("bam_str", "bam_test", "uam_str", "uam_rank", "prod", "disp")
)
}
\arguments{
\item{data}{Tidy-formatted data, with one token of a certain combination in each line. It should include at least columns for lemmas of the two slots, plus one for the document in which the combination appears.}

\item{lemma_names}{Column names for the lemmas of the two slots.}

\item{slot_names}{Abbreviations for the two slots.}

\item{doc_id}{Column name containing documents.}

\item{measures}{A list of abbreviations. See Notes below. If you just want the basic measures of each type (frequency, attraction, productivity, dispersion) and don't plan to use fancier meaures (e.g. based on LNRE models) or composite (e.g. lexical gravity), the default should work fine. #TODO: actually write the list,.}
}
\value{
A list. The first value, results, is usually all you want. You are also provided with several other views of the original data, including a list of combinations with frequencies by document, a term-document matrix, and a contingency table.
}
\description{
Get common co-occurrence measures.
}
